Create Insertion Sort Algorithm in Data Structures using Python
Sorting and Searching algorithms like insertion sort, selection sort, merge sort are very important to learn as a part of the preparation of your placements and also for internships because these form the base of any programmer. As a programmer, you have to deal with large amounts of data. Hence, we use sorting algorithms and one of them is an insertion sort. Now, the Insertion sort algorithm is a comparison-based sorting algorithm in which the sorted array is built by inserting one element at its correct position at a time.

Additionally, Insertion sort requires two arrays, one ordered, and one unordered. Firstly, we consider our sorted array with only one element that is a[0]. After that, on each repetition, we set a key from the unordered list and compare it with the elements present in the ordered list. Then we insert that key into a sorted position in the ordered list until there are no elements left in the unordered list. While understanding insertion sort one thing you need to keep in mind is “Insert an element from unsorted array to its correct position in the sorted array”.